Revert "Merging luis backend tweaks"

+++ backend/src/applications/applications.service.ts
@@ -6,21 +6,21 @@ export class ApplicationsService {
   constructor(private prisma: PrismaService) {}
 
   findAll(userId: string) {
-    return this.prisma.jobApplication.findMany({ where: { userId } });
+    return this.prisma.jobApplication.findMany({ where: { userId: Number(userId) } });
   }
 
   create(userId: string, data: any) {
-    return this.prisma.jobApplication.create({ data: { ...data, userId } });
+    return this.prisma.jobApplication.create({ data: { ...data, userId: Number(userId) } });
   }
 
   update(userId: string, id: string, data: any) {
     return this.prisma.jobApplication.updateMany({
-      where: { id: Number(id), userId },
+      where: { id: Number(id), userId: Number(userId)  },
       data,
     });
   }
 
   remove(userId: string, id: string) {
-    return this.prisma.jobApplication.deleteMany({ where: { id: Number(id), userId } });
+    return this.prisma.jobApplication.deleteMany({ where: { id: Number(id), userId: Number(userId) } });
   }
 }
